Unit-Tests: Fixes after refactoring
