+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/common/StringTests.kt
@@ -1,4 +1,4 @@
-package test
+package test.common
 
 import org.junit.Test
 import slatekit.common.repeatWith
