+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/entities/Entity_Database_Tests.kt
@@ -9,7 +9,7 @@ usage: Please refer to license on github for more info.
 </slate_header>
  */
 
-package slate.test
+package test.entities
 
 import org.junit.Test
 import java.time.*
